CASMPET-1248: Fix build
Fixes a couple of problems:

1) The GOPATH variable wasn't set so commands were executing
against the wrong files. The fix is to set the GOPATH variable to
the default ($HOME/go).

2) Old builds that had run and failed left the system in a broken
state where the go binary had been deleted. The fix for this is
to un-install and re-install the go packages so that the go binary
is in place.